修改LoggingReportK8sData结构

+++ logging.go
@@ -66,8 +66,8 @@ type LoggingReportK8sData struct {
 	P90          int64   `json:"p90,omitempty"`
 	P99          int64   `json:"p99,omitempty"`
 	P100         int64   `json:"p100,omitempty"`
-	CountService uint64  `json:"ct_svr,omitempty"`
-	CountPod     uint64  `json:"ct_pds,omitempty"`
+	CountService int     `json:"ct_svr,omitempty"`
+	CountPod     int     `json:"ct_pds,omitempty"`
 }
 
 type LoggingReportDockerData struct {
